requirements.txt
loguru<=0.7.2
matplotlib<=3.8.3
seaborn~=0.13.2
tabulate>=0.8.7,<=0.8.10
importlib-metadata==4.13.0
dask[distributed]
pydantic>=1.8.2,<=1.9.1
pandas<=2.2.1
PyYAML<=6.0.1
typing-extensions==4.5.0; python_version < '3.10'
typing-extensions>=4.15.0; python_version >= '3.10'
git+https://github.com/Cognitive-AI-Systems/pogema.git
